requirements.txt
aiohttp==3.7.3
alpha-vantage==2.3.1
appdirs==1.4.4
astroid==2.5.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bidict==0.21.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorgram.py==1.2.0
cryptography==3.4.7
cycler==0.10.0
distlib==0.3.2
django-timedeltafield==0.7.10
dnspython==2.1.0
dominate==2.5.2
email-validator==1.1.2
filelock==3.0.12
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-api-core==1.26.3
google-api-python-client==2.6.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.0.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
multidict==5.1.0
newsapi-python==0.2.6
numpy==1.20.1
oauthlib==3.1.1
optional-django==0.3.0
packaging==20.9
pandas==1.2.2
protobuf==3.17.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pylint==2.7.2
pyparsing==2.4.7
pyperclip==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-engineio==3.13.2
python-medium==1.1
python-socketio==4.6.0
pytz==2021.1
react==4.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
seaborn==0.11.1
six==1.15.0
soupsieve==2.2
spotipy==2.16.1
SQLAlchemy==1.4.18
threadpoolctl==2.1.0
timeseries==0.5.0
toml==0.10.2
twilio==6.51.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.4.7
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
yarl==1.6.3